+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [4.1.5] - 2020-07-30
 ### Added
 - Function to get Refresh URL's domain
+- Function to set and get Auth0's API path
 
 ## [4.1.4] - 2020-06-09
 ### Added